requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
nats-py==2.12.0
python-dotenv==1.0.1
pydantic==2.9.2
psycopg[binary]==3.1.18